requirements_dev.txt
-r requirements.txt

codecov==2.0.22
coloredlogs==14.0
coverage==5.0.4
factory_boy==2.8.1
flake8-blind-except==0.1.1
flake8-commas==2.0.0
flake8-docstrings==1.5.0
flake8-isort==2.9.1
flake8-pep3101==1.3.0
flake8-quotes==3.0.0
flake8-string-format==0.3.0
flake8==3.7.9
ipython==7.13.0
isort==4.3.21
mock==4.0.2
pydocstyle==5.0.2
pytest-cov==2.8.1
pytest-django==3.9.0
pytest==5.4.1
python-coveralls==2.9.3

